--- src/components/pages/index/OpenSource/Project/styles.module.css
@@ -1,73 +1,0 @@
-.projectContainer {
-  display: flex;
-}
-
-.project {
-  border-radius: .25rem;
-  border-top-color: var(--textSecondary);
-  border-top: .25rem solid;
-  box-shadow: 0 1px 4px var(--shadowColor);
-  color: var(--text);
-  flex: 1;
-  margin-bottom: 1rem;
-  padding: 2.5rem;
-  text-decoration: none;
-  transition: all 0.2s ease-out;
-}
-
-.starsContainer {
-  display: flex;
-  align-items: center;
-}
-
-.stars {
-  color: var(--languageJavaScript);
-  margin-right: .25rem;
-}
-
-@media(max-width: 768px) {
-  .project {
-    padding: 1.5rem;
-  }
-}
-
-.project:hover {
-  box-shadow: 0 .5rem 1rem var(--shadowColor);
-  color: inherit;
-  text-decoration: inherit;
-}
-
-@media (prefers-color-scheme: dark) {
-  .project {
-    background-color: var(--backgroundSurface);
-    box-shadow: none;
-  }
-
-  .project:hover {
-    background-color: var(--backgroundElevation);
-    box-shadow: none;
-  }
-}
-
-.projectName {
-  font-size: 1.25rem;
-  font-weight: 700;
-  margin: 0;
-}
-
-.project:global(.javascript) {
-  border-top-color: var(--languageJavaScript);
-}
-
-.project:global(.typescript) {
-  border-top-color: var(--languageTypeScript);
-}
-
-.project:global(.shell) {
-  border-top-color: var(--languageShell);
-}
-
-.starEmoji {
-  font-size: .95rem;
-  padding-right: .75rem;
-}
