+++ src/components/pages/index/Projects/__tests__/stubs.ts
